Migrate the component fees-table to Typescript
