+++ src/views/transactions/components/fees-table/fees-table.styles.ts
